Rename `stm32f::periph` namespace to `stm32f`
Also clang-formats everything.
