+++ firmware/mcal/stm32f/analog_output_dac.hpp
@@ -14,7 +14,7 @@
 #include "stm32f4xx_hal.h"
 #endif
 
-namespace mcal::stm32f::periph {
+namespace mcal::stm32f {
 
 class AnalogOutputDAC : public shared::periph::AnalogOutput {
 public:
@@ -34,4 +34,4 @@ private:
     uint32_t channel_;
 };
 
-}  // namespace mcal::stm32f::periph
+}  // namespace mcal::stm32f
