--- firmware/mcal/stm32f/periph/CMakeLists.txt
@@ -1,1 +1,0 @@
-target_sources(mcal-stm32f INTERFACE can.cc)
