+++ firmware/projects/Demo/AnalogDAC/platforms/nucleo-f767/bindings.cc
@@ -1,7 +1,7 @@
 #include "../../bindings.hpp"
 
 #include "dac.h"
-#include "mcal/stm32f/periph/analog_output_dac.hpp"
+#include "mcal/stm32f/analog_output_dac.hpp"
 
 extern "C" {
 /**
@@ -12,7 +12,7 @@ void SystemClock_Config();
 }
 
 namespace mcal {
-using namespace stm32f::periph;
+using namespace stm32f;
 
 AnalogOutputDAC analog_out_dac{&hdac, DAC_CHANNEL_1};
