+++ firmware/projects/Demo/CAN/Send/platforms/stm32f767/cubemx/board_config.ioc
@@ -72,7 +72,7 @@ ProjectManager.FreePins=false
 ProjectManager.HalAssertFull=false
 ProjectManager.HeapSize=0x200
 ProjectManager.KeepUserCode=false
-ProjectManager.LastFirmware=true
+ProjectManager.LastFirmware=false
 ProjectManager.LibraryCopy=1
 ProjectManager.MainLocation=Src
 ProjectManager.NoMain=true
