+++ firmware/projects/Demo/SoftwareReset/platforms/stm32f767/bindings.cc
@@ -2,17 +2,16 @@
 
 #include "gpio.h"
 #include "main.h"
-#include "mcal/stm32f/periph/gpio.hpp"
+#include "mcal/stm32f/gpio.hpp"
 #include "shared/periph/gpio.hpp"
 #include "stm32f7xx_hal.h"
 
 namespace mcal {
 
-stm32f::periph::DigitalOutput green_led{green_led_GPIO_Port, green_led_Pin};
-stm32f::periph::DigitalOutput red_led{red_led_GPIO_Port, red_led_Pin};
-stm32f::periph::DigitalOutput blue_led{blue_led_GPIO_Port, blue_led_Pin};
-stm32f::periph::DigitalInput reset_button{ResetButton_GPIO_Port,
-                                          ResetButton_Pin};
+stm32f::DigitalOutput green_led{green_led_GPIO_Port, green_led_Pin};
+stm32f::DigitalOutput red_led{red_led_GPIO_Port, red_led_Pin};
+stm32f::DigitalOutput blue_led{blue_led_GPIO_Port, blue_led_Pin};
+stm32f::DigitalInput reset_button{ResetButton_GPIO_Port, ResetButton_Pin};
 
 }  // namespace mcal
