+++ firmware/projects/FrontController/platforms/stm32-ev6/bindings.cc
@@ -4,9 +4,9 @@
 #include "can.h"
 #include "gpio.h"
 #include "main.h"
-#include "mcal/stm32f/periph/analog_input.hpp"
-#include "mcal/stm32f/periph/can.hpp"
-#include "mcal/stm32f/periph/gpio.hpp"
+#include "mcal/stm32f/analog_input.hpp"
+#include "mcal/stm32f/can.hpp"
+#include "mcal/stm32f/gpio.hpp"
 #include "stm32f7xx_hal.h"
 
 extern "C" {
@@ -16,7 +16,7 @@ void SystemClock_Config();
 }
 
 namespace mcal {
-using namespace stm32f::periph;
+using namespace stm32f;
 
 // =========== CAN =========================================
 CanBase veh_can_base{&hcan2};
